Title: 괄호 쌍 만들어주기, Time: 226ms, Memory: 22MB, Status: Passed - CodeTree

+++ 250322/괄호 쌍 만들어주기/pair-parentheses.java
@@ -9,7 +9,7 @@ public class Main {
 	public static void main(String[] args) {
 		input();
 		// x[i]는 i + 1부터 끝까지 ))가 몇개있는지?
-		int[] x = new int[str.length()];
+		long[] x = new long[str.length()];
 
 		for (int i = x.length - 1; i >= 0; i--) {
 			// 맨처음에는 연속해서 ))을 셀 수 없으므로 넘긴다.
@@ -20,7 +20,7 @@ public class Main {
 			x[i] = "))".equals(temp) ? x[i + 1] + 1 : x[i + 1];
 		}
 
-		int ans = 0;
+		long ans = 0;
 		for (int i = 0; i < x.length - 1; i++) {
 			String temp = str.substring(i, i + 2);
 			if ("((".equals(temp)) {
